refactor: align save upload compression with review feedback
